refactor(strategy): 장기판 초기 배치 전략 패키지를 board 하위로 이동
- 장기판 조립 및 기물 배치 전략 객체가 `Board` 객체의 생성 주기를 담당하므로, 응집도를 높이기 위해 strategy 패키지의 위치를 board 패키지의 하위로 이동
- 패키지 구조 변경에 따른 import 경로 일괄 업데이트
