+++ src/main/java/janggi/controller/JanggiFlow.java
@@ -6,9 +6,9 @@ import janggi.domain.Side;
 import janggi.domain.board.Board;
 import janggi.domain.piece.Piece;
 import janggi.domain.piece.PieceType;
-import janggi.domain.strategy.ArrangementOption;
-import janggi.domain.strategy.ArrangementStrategy;
-import janggi.domain.strategy.BoardAssembler;
+import janggi.domain.board.strategy.ArrangementOption;
+import janggi.domain.board.strategy.ArrangementStrategy;
+import janggi.domain.board.strategy.BoardAssembler;
 import janggi.view.ApplicationView;
 import janggi.view.label.ArrangementStrategyLabel;
 import java.util.List;
