+++ src/main/java/janggi/domain/board/strategy/ArrangementStrategy.java
@@ -1,4 +1,4 @@
-package janggi.domain.strategy;
+package janggi.domain.board.strategy;
 
 import janggi.domain.piece.Piece;
 import janggi.domain.piece.PieceFactory;
