+++ src/main/java/janggi/domain/board/strategy/MaSangArrangementStrategy.java
@@ -1,4 +1,4 @@
-package janggi.domain.strategy;
+package janggi.domain.board.strategy;
 
 import janggi.domain.Side;
 import janggi.domain.piece.Piece;
