+++ src/test/java/janggi/domain/board/strategy/MaSangArrangementStrategyTest.java
@@ -1,4 +1,4 @@
-package janggi.strategy;
+package janggi.domain.board.strategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,7 +7,7 @@ import janggi.domain.board.GungSeong;
 import janggi.domain.piece.Piece;
 import janggi.domain.piece.PieceFactory;
 import janggi.domain.piece.PieceType;
-import janggi.domain.strategy.MaSangArrangementStrategy;
+import janggi.domain.board.strategy.MaSangArrangementStrategy;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
