+++ src/test/java/janggi/support/TestArrangementStrategy.java
@@ -3,7 +3,7 @@ package janggi.support;
 import janggi.domain.Location;
 import janggi.domain.piece.Piece;
 import janggi.domain.piece.PieceFactory;
-import janggi.domain.strategy.ArrangementStrategy;
+import janggi.domain.board.strategy.ArrangementStrategy;
 import java.util.Map;
 
 public class TestArrangementStrategy implements ArrangementStrategy {
